requirements.txt
numpy==1.18.5
Pillow>=9.0.1
PyOpenGL==3.1.5
PyGLM==2.2.0
pyserial==3.5
six==1.14.0
wxPython>=4.1.0
PyDispatcher==2.0.5
PyWavefront>=1.3.2
mprop==0.16.0
platformdirs>=4.3.6
exif>=1.6.0
PyExifTool>=0.5.6
